+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.3
+- Changed return string in solver.check and solver.model to be more in line with C library
+- Made tests from examples (beware:including smelly tests!)
+
 ## 1.0.2
 - Removed linter for generated bindings and fixed other small problems